# go/model_container.go
/*
 * Inventory API
 *
 * This is a sample inventory management server based on the OpenAPI 3.0 specification
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Container struct {

	Id int64 `json:"id,omitempty"`

	ItemCount int64 `json:"itemCount,omitempty"`
}