# go/model_item.go
/*
 * Inventory API
 *
 * This is a sample inventory management server based on the OpenAPI 3.0 specification
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Item struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	CategoryId int64 `json:"categoryId,omitempty"`

	ContainerId int64 `json:"containerId"`
}